fix: restructure split modules to use proper directory layout
- Move parser.rs to parser/mod.rs with tests.rs in same directory
- Move store.rs to store/mod.rs with tests.rs in same directory
- Add clippy allow for similar_names in parser tests
- Fix formatting issues in extracted test files
