+++ crates/breez-sdk/common/src/input/parser/mod.rs
@@ -883,6 +883,5 @@ fn parse_silent_payment_address(
     None
 }
 
-
 #[cfg(test)]
 mod tests;
